Split into files based on functionality
